mysite/requirements.txt
asgiref==3.2.3
defusedxml==0.6.0
Django==2.2.1
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-xml==1.4.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
